+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+Development:
+  - add more outpu to "deposit verify" to explain operation
 1.7.1:
   - fix "store not set" issue
 1.7.0: